Views should be using cache busting as well
Caching (and partially cache busting) of CSS and JS files are handled by the ClientDependency framework, while HTML files are cached by the browser, and not properly "uncached" when modified. By specifying the package version in the query string, the browser will automatically load the updated views when a new version is installed.